requirements.txt
Django==5.1.1
djangorestframework==3.15.2
ultralytics==8.3.0
opencv-python==4.10.0.84
pymongo==4.9.0
dnspython==2.6.1
